Common PolicyCreateForm
