+++ portals/publisher/src/main/webapp/source/src/app/components/GlobalPolicies/Policies/PolicyForm/PolicyCreateForm.tsx
@@ -17,127 +17,14 @@
  */
 
 import React, { FC, useEffect, useReducer, useState } from 'react';
-import { CircularProgress, makeStyles, Theme } from '@material-ui/core';
-import Box from '@material-ui/core/Box';
-import Divider from '@material-ui/core/Divider';
-import Button from '@material-ui/core/Button';
-import Paper from '@material-ui/core/Paper';
-import { FormattedMessage } from 'react-intl';
-import { isRestricted } from 'AppData/AuthManager';
 import CONSTS from 'AppData/Constants';
+import PolicyCreateFormShared from 'AppComponents/Shared/PoliciesUI/PolicyCreateForm';
 import type { CreatePolicySpec } from '../Types';
 import type { NewPolicyState, PolicyAttribute } from './Types';
 import PolicyAttributes from './PolicyAttributes';
 import uuidv4 from '../Utils';
 import GeneralDetails from './GeneralDetails';
 import SourceDetails from './SourceDetails';
-
-/** Shared UI Component */
-const useStyles = makeStyles((theme: Theme) => ({
-    root: {
-        flexGrow: 1,
-        marginTop: 10,
-        display: 'flex',
-        flexDirection: 'column',
-        padding: 20,
-    },
-    cancelBtn: {
-        marginLeft: theme.spacing(1),
-    },
-}));
-
-interface PolicyCreateFormSharedProps {
-    synapsePolicyDefinitionFile: any[];
-    setSynapsePolicyDefinitionFile: React.Dispatch<React.SetStateAction<any[]>>;
-    ccPolicyDefinitionFile: any[];
-    setCcPolicyDefinitionFile: React.Dispatch<React.SetStateAction<any[]>>;
-    onCancel: () => void;
-    saving: boolean;
-    state: any;
-    dispatch: any;
-    onPolicySave: () => void;
-    isFormDisabled: boolean;
-}
-
-const PolicyCreateFormShared: FC<PolicyCreateFormSharedProps> = ({
-    synapsePolicyDefinitionFile,
-    setSynapsePolicyDefinitionFile,
-    ccPolicyDefinitionFile,
-    setCcPolicyDefinitionFile,
-    onCancel,
-    saving,
-    state,
-    dispatch,
-    onPolicySave,
-    isFormDisabled
-}) => {
-    const classes = useStyles();
-
-    return (
-        <Paper elevation={0} className={classes.root} data-testid='create-policy-form'>
-            {/* General details of policy */}
-            <GeneralDetails
-                displayName={state.displayName}
-                version={state.version}
-                description={state.description}
-                applicableFlows={state.applicableFlows}
-                supportedApiTypes={state.supportedApiTypes}
-                dispatch={dispatch}
-                isViewMode={false}
-            />
-            <Divider light />
-            {/* Gateway specific details of policy */}
-            <SourceDetails
-                supportedGateways={state.supportedGateways}
-                synapsePolicyDefinitionFile={synapsePolicyDefinitionFile}
-                setSynapsePolicyDefinitionFile={setSynapsePolicyDefinitionFile}
-                ccPolicyDefinitionFile={ccPolicyDefinitionFile}
-                setCcPolicyDefinitionFile={setCcPolicyDefinitionFile}
-                dispatch={dispatch}
-            />
-            <Divider light />
-            {/* Attributes of policy */}
-            <PolicyAttributes
-                policyAttributes={state.policyAttributes}
-                dispatch={dispatch}
-                isViewMode={false}
-            />
-            <Box data-testid='policy-add-btn-panel'>
-                <Button
-                    variant='contained'
-                    color='primary'
-                    onClick={onPolicySave}
-                    data-testid='policy-create-save-btn'
-                    disabled={
-                        isRestricted([
-                            'apim:api_create',
-                            'apim:api_manage',
-                            'apim:mediation_policy_create',
-                            'apim:mediation_policy_manage',
-                            'apim:api_mediation_policy_manage',
-                        ]) || isFormDisabled
-                    }
-                >
-                    {saving ? (
-                        <CircularProgress size={16} />
-                    ) : (
-                        <FormattedMessage
-                            id='Apis.Details.Policies.PolicyForm.PolicyCreateForm.policy.save'
-                            defaultMessage='Save'
-                        />
-                    )}
-                </Button>
-                <Button className={classes.cancelBtn} onClick={onCancel}>
-                    <FormattedMessage
-                        id='Apis.Details.Policies.PolicyForm.PolicyCreateForm.policy.cancel'
-                        defaultMessage='Cancel'
-                    />
-                </Button>
-            </Box>
-        </Paper> 
-    );
-}
-/** Shared UI Component Ends */
 
 export const ACTIONS = {
     UPDATE_POLICY_METADATA: 'updatePolicyMetadata',
@@ -382,6 +269,10 @@ const PolicyCreateForm: FC<PolicyCreateFormProps> = ({
             dispatch={dispatch}
             onPolicySave={onPolicySave}
             isFormDisabled={isFormDisabled}
+            PolicyAttributes={PolicyAttributes}
+            GeneralDetails={GeneralDetails}
+            SourceDetails={SourceDetails}
+            isRestricted={false}
         />
     );
 };
